+++ Source/ModuleKISInventory.cs
@@ -7,6 +7,7 @@ using KIS.GUIUtils;
 using KSPDev.ConfigUtils;
 using KSPDev.GUIUtils;
 using KSPDev.LogUtils;
+using KSPDev.ModelUtils;
 using KSPDev.KSPInterfaces;
 using KSPDev.PartUtils;
 using System;
@@ -473,6 +474,8 @@ public class ModuleKISInventory : PartModule,
   public string invName = "";
   [KSPField(isPersistant = true)]
   public bool helmetEquipped = true;
+  [KSPField]
+  public InventoryType invType = InventoryType.Container;
   #endregion
 
   #region Global settings
@@ -529,7 +532,6 @@ public class ModuleKISInventory : PartModule,
   public string openGuiName;
   public float totalVolume = 0;
   public int podSeat = -1;
-  public InventoryType invType = InventoryType.Container;
   public enum InventoryType {
     Container,
     Pod,
@@ -1330,24 +1332,20 @@ public class ModuleKISInventory : PartModule,
       }
     }
 
-    //Disable helmet and visor
-    var skmrs =
-        new List<SkinnedMeshRenderer>(part.GetComponentsInChildren<SkinnedMeshRenderer>());
-    foreach (var skmr in skmrs) {
-      if (skmr.name == "helmet" || skmr.name == "visor") {
-        skmr.GetComponent<Renderer>().enabled = active;
-        helmetEquipped = active;
-      }
-    }
-
-    //Disable flares and light
-    var lights = new List<Light>(part.GetComponentsInChildren<Light>(true) as Light[]);
-    foreach (var light in lights) {
-      if (light.name == "headlamp") {
-        light.enabled = active;
-        light.transform.Find("flare1").GetComponent<Renderer>().enabled = active;
-        light.transform.Find("flare2").GetComponent<Renderer>().enabled = active;
-      }
+    // Disable helmet and visor.
+    var helmet = Hierarchy.FindTransformByPath(part.transform, "**/helmet*");
+    if (helmet != null) {
+      HostedDebugLog.Fine(this, "Disable helmet renderers and lights on {0}", part);
+      helmet.GetComponentsInChildren<Renderer>(includeInactive: true)
+          .ToList()
+          .ForEach(r => r.enabled = active);
+      part.transform.GetComponentsInChildren<Light>(includeInactive: true)
+          .Where(l => l.name == "headlamp")
+          .ToList()
+          .ForEach(l => l.enabled = active);
+      helmetEquipped = active;
+    } else {
+      HostedDebugLog.Error(this, "Cannot find the helmet model on {0}!", part);
     }
 
     return true;